+++ Game/src/main/java/com/example/game/Player.java
@@ -30,7 +30,7 @@ public class Player implements Serializable {
 
     }
 
-    public int getSpeed(){
+    public double getSpeed(){
         return hero.getSpeed();
     }
